refactor base resource and collection
