+++ tamr_unify_client/models/attribute/resource.py
@@ -1,5 +1,5 @@
+from tamr_unify_client.base_resource import BaseResource
 from tamr_unify_client.models.attribute.type import AttributeType
-from tamr_unify_client.models.base_resource import BaseResource
 
 
 class Attribute(BaseResource):
