+++ tamr_unify_client/models/binning_model.py
@@ -1,6 +1,6 @@
 import json
 
-from tamr_unify_client.models.base_resource import BaseResource
+from tamr_unify_client.base_resource import BaseResource
 
 
 class BinningModel(BaseResource):
