+++ tamr_unify_client/models/category/resource.py
@@ -1,4 +1,4 @@
-from tamr_unify_client.models.base_resource import BaseResource
+from tamr_unify_client.base_resource import BaseResource
 
 
 class Category(BaseResource):
