+++ tamr_unify_client/models/project/cluster_configuration.py
@@ -1,4 +1,4 @@
-from tamr_unify_client.models.base_resource import BaseResource
+from tamr_unify_client.base_resource import BaseResource
 
 
 class PublishedClustersConfiguration(BaseResource):
